+++ src/components/core/Footer/styled.js
